moved allowance finding out of PatternPath's methods and into a static helper class. Added a check to prevent computing the allowance of an allowance
